Configured base URL for dev and production.
When deployed to GitHub Pages, there will be a base URL of "/snippety/" because that is the name of the repo.

The base URL is not needed for dev however having the same base URL should help catch any routing issues that might have only been spotted when deploying to production.